Authentication.login {SciServer}	R Documentation

User login.

Description

Logs the user into SciServer and returns the authentication token. This function is useful when SciScript-R library methods are executed outside the SciServer-Compute environment. In this case, the session authentication token does not exist (and therefore can't be automatically recognized), so the user has to use Authentication.login in order to log into SciServer manually and get the authentication token. Authentication.login also sets the token value in the Sys environment variable "sciservertoken" and as the variable Authentication.token.

Usage

Authentication.login(UserName, Password)

Arguments

UserName
name of the user.

Password
password of the user.

Value

Returns the authentication token (string). Throws an exception if the HTTP request to the Authentication URL returns an error.

Author(s)

Gerard Lemson, Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://portal.sciserver.org/login-portal/
http://www.github.com/sciserver/SciScript-R

See Also

Authentication.getKeystoneUserWithToken, Authentication.getToken, Authentication.setToken, Authentication.token.

Examples

## Not run: Authentication.login("MyName","MyPassword") 

[Package SciServer version 1.10.0 Index]