Authentication.setToken {SciServer}	R Documentation

Set token.

Description

Sets the SciServer authentication token of the user as the Sys environmental variable, as well as in the variable Authentication.token.

Usage

Authentication.setToken(token)

Arguments

token
Sciserver's authentication token for the user

Author(s)

Gerard Lemson, Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://portal.sciserver.org/login-portal/
http://www.github.com/sciserver/SciScript-R

See Also

Authentication.getKeystoneUserWithToken, Authentication.login, Authentication.getToken, Authentication.token.

Examples

## Not run: 
Authentication.setToken("MyToken")
## End(Not run)

[Package SciServer version 1.10.0 Index]