Authentication.token {SciServer}	R Documentation

Stores the SciServer user authentication token.

Description

String variable that stores the authentication token for the user's session. It has an initially a NULL value, and gets set directly by running Authentication.setToken, and indirectly by running Authentication.login or Authentication.getToken.

Usage

Authentication.token

Value

Stores the authentication token (string). If never set, it stores a NULL value.

Author(s)

Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://portal.sciserver.org/login-portal/
http://www.github.com/sciserver/SciScript-R

See Also

Authentication.getKeystoneUserWithToken, Authentication.getToken, Authentication.setToken.

Examples

## Not run: myToken = Authentication.token 

[Package SciServer version 1.10.0 Index]