CasJobs.executeQuery {SciServer}	R Documentation

Synchronous SQL query.

Description

Executes a synchronous SQL query in a CasJobs database context.

Usage

CasJobs.executeQuery(sql, context="MyDB", format="dataframe")

Arguments

sql
sql query

context
database context

format
output format. Choose between "list" (list object), "dataframe" (data.frame object), "csv" (csv string), "json" (json string), "fits" (raw binary data containing a fits file) and "raw" (raw binary data containing a fits file).

Value

Returns the results table from the query in different output formats defined by the input parameter 'format'. Throws an exception if the HTTP request to the CasJobs API returns an error

Author(s)

Gerard Lemson, Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://skyserver.sdss.org/casjobs
http://www.github.com/sciserver/SciScript-R

See Also

CasJobs.submitJob, CasJobs.getTables, SkyServer.sqlSearch

Examples

table = CasJobs.executeQuery(sql="select 1 as foo, 2 as bar",context="MyDB")

[Package SciServer version 1.10.0 Index]