CasJobs.getJobStatus {SciServer}	R Documentation

Get status of job.

Description

Shows the status of a job submitted to CasJobs.

Usage

CasJobs.getJobStatus(jobId)

Arguments

jobId
id of the job

Value

Returns a list containing the job status and related metadata. The "Status" field can be equal to 0 (Ready), 1 (Started), 2 (Canceling), 3(Canceled), 4 (Failed) or 5 (Finished). If jobid is the empty string, then returns a list with the statuses of all previous jobs. Throws an exception if the user is not logged into SciServer (use Authentication.login for that purpose). Throws an exception if the HTTP request to the CasJobs API returns an error.

Author(s)

Gerard Lemson, Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://skyserver.sdss.org/casjobs
http://www.github.com/sciserver/SciScript-R

See Also

CasJobs.submitJob, CasJobs.waitForJob, CasJobs.cancelJob.

Examples

## Not run: status = CasJobs.getJobStatus(CasJobs.submitJob("select 1"))

[Package SciServer version 2.0.0 Index]