CasJobs.getTables {SciServer}	R Documentation

Get tables descriptions.

Description

Gets the names, size and creation date of all tables in a database context that the user has access to.

Usage

CasJobs.getTables(context)

Arguments

context
database context

Value

Returns a 'list', which includes the names, size and creation date of all tables. Throws an exception if the user is not logged into SciServer (use Authentication.login for that purpose). Throws an exception if the HTTP request to the CasJobs API returns an error.

Author(s)

Gerard Lemson, Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://skyserver.sdss.org/casjobs
http://www.github.com/sciserver/SciScript-R

See Also

CasJobs.getSchemaName.

Examples

## Not run: tables = CasJobs.getTables(context="MyDB")

[Package SciServer version 1.10.0 Index]