CasJobs.waitForJob {SciServer}	R Documentation

Wait for job status.

Description

Queries the job status from casjobs every 2 seconds and waits for the casjobs job to return a status of 3, 4, or 5 (Cancelled, Failed or Finished, respectively).

Usage

CasJobs.waitForJob(jobId, verbose=TRUE)

Arguments

jobId
id of job (integer).

verbose
message printing option.

If TRUE, will print "wait" messages on the screen while the job is not done. If FALSE, will suppress printing messages on the screen.

Value

Returns a list containing the job status and related metadata. The "Status" field can be equal to 0 (Ready), 1 (Started), 2 (Canceling), 3(Canceled), 4 (Failed) or 5 (Finished). Throws an exception if the user is not logged into SciServer (use Authentication.login for that purpose). Throws an exception if the HTTP request to the CasJobs API returns an error.

Author(s)

Gerard Lemson, Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://skyserver.sdss.org/casjobs
http://www.github.com/sciserver/SciScript-R

See Also

CasJobs.submitJob, CasJobs.getJobStatus, CasJobs.cancelJob.

Examples

CasJobs.waitForJob(CasJobs.submitJob("select 1"))

[Package SciServer version 2.0.0 Index]