CasJobs.writeFitsFileFromQuery {SciServer}	R Documentation

Write Fits file.

Description

Executes a quick CasJobs query and writes the result to a local Fits file (http://www.stsci.edu/institute/software_hardware/pyfits).

Usage

CasJobs.writeFitsFileFromQuery(fileName, queryString, context="MyDB")

Arguments

fileName
path to the local Fits file to be created (string).

queryString
sql query (string)

context
database context (string)

Value

Returns True if the fits file was created successfully. Throws an exception if the user is not logged into SciServer (use Authentication.login for that purpose). Throws an exception if the HTTP request to the CasJobs API returns an error.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://skyserver.sdss.org/casjobs
http://www.github.com/sciserver/SciScript-R

See Also

CasJobs.uploadCSVDataToTable.

Examples

response = CasJobs.writeFitsFileFromQuery(fileName="MyFile.fits", queryString="select 1 as foo,2 as bar", context="MyDB")

[Package SciServer version 2.0.0 Index]