Files.createDir {SciServer}	R Documentation

Create a directory.

Description

Create a directory or folder.

Usage

Jobs.createDir(fileService, path, quiet=TRUE)

Arguments

fileService
name of fileService (string), or object (list) that defines a file service. A list of these kind of objects available to the user is returned by the function Files.getFileServices().

path
path (in the remote file service) to the directory (string), starting from the root volume level. Example: rootVolume/userVolumeOwner/userVolume/directory

quiet
if set to FALSE, will throw an error if the directory already exists. If TRUE, won't throw an error.

Value

void

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://apps.sciserver.org/dashboard/files
http://www.github.com/sciserver/SciScript-R

See Also

Files.getFileServices, Files.getFileServiceFromName, Files.delete, Files.upload, Files.download, Files.dirList, Files.createUserVolume

Examples

fileServices = Files.getFileServices(); Files.createDir(fileServices[[1]], "rootVolume/userVolumeOwner/userVolume/newDirectory");

[Package SciServer version 2.0.0 Index]