Files.createUserVolume {SciServer}	R Documentation

Create a user volume.

Description

Create a user volume.

Usage

Jobs.createUserVolume(fileService, path, quiet=TRUE)

Arguments

fileService
name of fileService (string), or object (list) that defines a file service. A list of these kind of objects available to the user is returned by the function Files.getFileServices().

path
path (in the remote file service) to the user volume (string), starting from the root volume level. Example: rootVolume/userVolumeOwner/userVolume

quiet
if set to FALSE, will throw an error if the User Volume already exists. If TRUE, won't throw an error.

Value

void

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://apps.sciserver.org/dashboard/files
http://www.github.com/sciserver/SciScript-R

See Also

Files.getFileServices, Files.getFileServiceFromName, Files.delete, Files.upload, Files.download, Files.dirList, Files.deleteUserVolume

Examples

fileServices = Files.getFileServices(); Files.createUserVolume(fileServices[[1]], "rootVolume/userVolumeOwner/userVolume");

[Package SciServer version 2.0.0 Index]