Files.delete {SciServer}	R Documentation

Delete file or folder

Description

Deletes a directory or file in the File System.

Usage

Jobs.delete(fileService, path, quiet=TRUE)

Arguments

fileService
name of fileService (string), or object (dictionary) that defines a file service. A list of these kind of objects available to the user is returned by the function Files.getFileServices().

path
String defining the path (in the remote fileService) of the file or directory to be deleted, starting from the root volume level. Example: rootVolume/userVolumeOwner/userVolume/fileToBeDeleted.txt.

quiet
If set to FALSE, it will throw an error if the file does not exist. If set to TRUE. it will not throw an error.

Value

void

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://apps.sciserver.org/dashboard/files
http://www.github.com/sciserver/SciScript-R

See Also

Files.getFileServices, Files.getFileServiceFromName, Files.move, Files.createDir, Files.upload, Files.download, Files.createUserVolume

Examples

fileServices = Files.getFileServices(); Files.delete(fileServices[[1]],"rootVolume/userVolumeOwner/userVolume/fileToBeDeleted.txt");

[Package SciServer version 2.0.0 Index]