Files.dirList {SciServer}	R Documentation

List directory contents

Description

Lists the contents of a directory.

Usage

Jobs.dirList(fileService, path, level=1, options="")

Arguments

fileService
name of fileService (string), or object (list) that defines a file service. A list of these kind of objects available to the user is returned by the function Files.getFileServices().

path
String defining the path (in the remote file service) of the directory to be listed, starting from the root volume level. Example: rootVolume/userVolumeOwner/userVolume/directoryToBeListed.

level
amount (int) of listed directory levels that are below or at the same level to that of the relativePath.

options
string specifiying the file or directory patterns to match, based on the glob syntax

Value

list containing the directory listing.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://apps.sciserver.org/dashboard/files
http://www.github.com/sciserver/SciScript-R

See Also

Files.getFileServices, Files.getFileServiceFromName, Files.delete, Files.createDir, Files.upload, Files.download, Files.createUserVolume

Examples

fileServices = Files.getFileServices(); dirs = Files.dirList(fileService=fileServices[[1]], path="rootVolume/userVolumeOwner/userVolume/directoryToBeListed", level=4)

[Package SciServer version 2.0.0 Index]