Files.getFileServiceFromName {SciServer}	R Documentation

Get FileService from name

Description

Returns a FileService object, given its registered name.

Usage

Jobs.getFileServiceFromName(fileServiceName, fileServices=NULL, verbose=TRUE)

Arguments

fileServiceName
name of the FileService, as shown within the results of Files.getFileServices()

fileServices
a list of FileService objects (lists), as returned by Files.getFileServices(). If not set, then an extra internal call to Jobs.getFileServices() is made.

verbose
boolean parameter defining whether warnings will be printed (set to TRUE) or not (set to FALSE).

Value

a FileService object (list) that defines a FileService. A list of these kind of objects available to the user is returned by the function Jobs.getFileServices(). If no fileService can be found, then returns NULL.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://apps.sciserver.org/dashboard/files
http://www.github.com/sciserver/SciScript-R

See Also

Files.getFileServices

Examples

fileService = Files.getFileServiceFromName('FileServiceAtJHU');

[Package SciServer version 2.0.0 Index]