Files.getFileServicesNames {SciServer}	R Documentation

Get FileServices Names

Description

Returns the names and description of the fileServices available to the user.

Usage

Jobs.getFileServicesNames(fileServices=NULL, verbose=TRUE)

Arguments

fileServices
a list of FileService objects, as returned by Files.getFileServices(). If not set, then an extra internal call to Jobs.getFileServices() is made.

verbose
boolean parameter defining whether warnings will be printed (set to TRUE) or not (set to FALSE).

Value

a list of lists, each of which having the name and description of a file service available to the user.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://apps.sciserver.org/dashboard/files
http://www.github.com/sciserver/SciScript-R

See Also

Files.getFileServices

Examples

fileServiceNames = Files.getFileServicesNames()

[Package SciServer version 2.0.0 Index]