Files.getFileServices {SciServer}	R Documentation

Get FileServices

Description

Gets the definitions of file services that a user is able to access. A FileService represents a file system that contains root volumes accessible to the user for public/private data storage. Within each rootVolume, users can create sharable userVolumes for storing files.

Usage

Jobs.getFileServices(verbose=TRUE)

Arguments

verbose
boolean parameter defining whether warnings will be printed (set to TRUE) or not (set to FALSE).

Value

list that contains lists, each of which representing the description of a FileService that the user is able to access.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://apps.sciserver.org/dashboard/files
http://www.github.com/sciserver/SciScript-R

See Also

Files.getFileServiceFromName

Examples

fileServices = Files.getFileServices()

[Package SciServer version 2.0.0 Index]