Files.getRootVolumesInfo {SciServer}	R Documentation

Get RootVolumes info

Description

Gets the names and descriptions of root volumes available to the user in a particular FileService.

Usage

Jobs.getRootVolumesInfo(fileService, verbose=TRUE)

Arguments

fileService
name of fileService (string), or object (list) that defines a file service. A list of these kind of objects available to the user is returned by the function Files.getFileServices().

verbose
boolean parameter defining whether warnings will be printed (set to TRUE) or not (set to FALSE).

Value

list of lists, each of which containing the name and description of a root volume.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://apps.sciserver.org/dashboard/files
http://www.github.com/sciserver/SciScript-R

See Also

Files.getFileServices, Files.getUserVolumesInfo

Examples

fileServices = Files.getFileServices(); rootVolumesInfo = getRootVolumesInfo(fileServices[[1]])

[Package SciServer version 2.0.0 Index]