Files.getUserVolumesInfo {SciServer}	R Documentation

Get UserVolumes info

Description

Gets the names and definitions of the UserVolumes available in a particular FileService

Usage

Jobs.getUserVolumesInfo(fileService, rootVolumeName=NULL, verbose=TRUE)

Arguments

fileService
name of fileService (string), or object (list) that defines a file service. A list of these kind of objects available to the user is returned by the function Files.getFileServices().

rootVolumeName
name of root Volume (string) for which the user volumes are fetched. If set to NULL, then user volumes in all root folders are fetched.

verbose
boolean parameter defining whether warnings will be printed (set to TRUE) or not (set to FALSE).

Value

list of lists, each of which containing the name and description of a user volume that a user is able to access.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://apps.sciserver.org/dashboard/files
http://www.github.com/sciserver/SciScript-R

See Also

Files.getFileServices, Files.getRootVolumesInfo

Examples

fileServices = Files.getFileServices(); userVolumesInfo = Files.getUserVolumesInfo(fileServices[[1]])

[Package SciServer version 2.0.0 Index]