Files.shareUserVolume {SciServer}	R Documentation

Share a user volume

Description

Shares a user volume with another user or group

Usage

Jobs.shareUserVolume(fileService, path, sharedWith, allowedActions, type="USER")

Arguments

fileService
name of fileService (string), or object (dictionary) that defines a file service. A list of these kind of objects available to the user is returned by the function Files.getFileServices().

path
String defining the path (in the remote fileService) of the user volume to be shared, starting from the root volume level. Example: rootVolume/userVolumeOwner/userVolume.

sharedWith
name (string) of user or group that the user volume is shared with.

allowedActions
list of strings defining actions the user or group is allowed to do with respect to the shared user volume. E.g.: list("read"). Type of actions are "read", "write", "grant" (the user or group can also share the user volume with another user or group) and "delete" (ability to delete the user volume , use with care).

type
type (string) of the entity defined by the "sharedWith" parameter. Can be set to "USER" or "GROUP".

Value

void

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://apps.sciserver.org/dashboard/files
http://www.github.com/sciserver/SciScript-R

See Also

Files.getFileServices, Files.getFileServiceFromName, Files.move, Files.createDir, Files.upload, Files.download, Files.createUserVolume

Examples

fileServices = Files.shareUserVolume(); Files.shareUserVolume(fileService[[1]],"rootVolume/userVolumeOwner/userVolume",sharedWith="userName", allowedActions=list("read"), type="USER")

[Package SciServer version 2.0.0 Index]