Jobs.getDockerComputeDomainFromName {SciServer}	R Documentation

Get a DockerComputeDomain

Description

Returns the names of the docker compute domains available to the user.

Usage

Jobs.getDockerComputeDomainFromName(dockerComputeDomainName, dockerComputeDomains = NULL)

Arguments

dockerComputeDomainName
name of the DockerComputeDomainName, as shown within the results of Jobs.getDockerComputeDomains()

dockerComputeDomains
a list of dockerComputeDomain objects (dictionaries), as returned by Jobs.getDockerComputeDomains(). If not set, then an internal call to Jobs.getDockerComputeDomains() is made.

Value

a DockerComputeDomain object (list) that defines a Docker compute domain. A list of these kind of objects available to the user is returned by the function Jobs.getDockerComputeDomains().

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://www.github.com/sciserver/SciScript-R

See Also

Jobs.getDockerComputeDomains.

Examples

dockerComputeDomain = Jobs.getDockerComputeDomainFromName("dockerComputeDomainAtJHU");

[Package SciServer version 2.0.0 Index]