Jobs.getDockerComputeDomains {SciServer}	R Documentation

Get Docker compute domains

Description

Gets a list of all registered Docker compute domains that the user has access to.

Usage

Jobs.getDockerComputeDomains()

Value

a list of objects, each one containing the definition of a Docker compute domain.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://www.github.com/sciserver/SciScript-R

See Also

Jobs.submitShellCommandJob, Jobs.getJobStatus, Jobs.getRDBComputeDomains, Jobs.cancelJob.

Examples

dockerComputeDomains = Jobs.getDockerComputeDomains()

[Package SciServer version 2.0.0 Index]