Jobs.getDockerComputeDomainsNames {SciServer}	R Documentation

Get docker compute domains names.

Description

Returns the names of the docker compute domains available to the user.

Usage

Jobs.getDockerComputeDomainsNames(dockerComputeDomains=NULL)

Arguments

dockerComputeDomains
a list of dockerComputeDomain objects, as returned by Jobs.getDockerComputeDomains(). If not set, then an extra internal call to Jobs.getDockerComputeDomains() is made.

Value

a list of strings, each being the name of a docker compute domain available to the user.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://www.github.com/sciserver/SciScript-R

See Also

Jobs.getDockerComputeDomains.

Examples

dockerComputeDomainsNames = Jobs.getDockerComputeDomainsNames()

[Package SciServer version 2.0.0 Index]