Jobs.getJobStatus {SciServer}	R Documentation

Get job status

Description

Gets a list with the job status as an integer value, together with its semantic meaning. The integer value is a power of 2, that is, 1:PENDING, 2:QUEUED, 4:ACCEPTED, 8:STARTED, 16:FINISHED, 32:SUCCESS, 64:ERROR and 128:CANCELED

Usage

Jobs.getJobDescription(jobId)

Arguments

jobId
Id of job (integer)

Value

list containing the integer value of the job status, as well as its semantic meaning.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://www.github.com/sciserver/SciScript-R

See Also

Jobs.submitShellCommandJob,Jobs.getJobStatus,Jobs.getDockerComputeDomains,Jobs.cancelJob

Examples

job = Jobs.getJobStatus(123);

[Package SciServer version 2.0.0 Index]