Jobs.getJobsList {SciServer}	R Documentation

Get Jobs List

Description

Gets the list of Jobs submitted by the user.

Usage

Jobs.getJobsList(top=10, open=NULL, start=NULL, end=NULL, type="all")

Arguments

top
top number of jobs (integer) returned. If top=NULL, then all jobs are returned.

open
Boolean variable. If set to TRUE, then only returns jobs that have not finished executing and wrapped up (status <= FINISHED). If set to FALSE then only returnes jobs that are still running. If set to NULL, then returns both finished and unfinished jobs.

start
The earliest date (inclusive) to search for jobs, in string format yyyy-MM-dd hh:mm:ss.SSS. If set to NULL, then there is no lower bound on date.

end
The latest date (inclusive) to search for jobs, in string format yyyy-MM-dd hh:mm:ss.SSS. If set to NULL, then there is no upper bound on date.

type
type (string) of jobs returned. Can take values of 'rdb' (for returning only relational database jobs), 'docker' (for returning only Docker jobs) and 'all' (all job types are returned).

Value

a list containing the definitions of submitted jobs.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://www.github.com/sciserver/SciScript-R

See Also

Jobs.submitNotebookJob,Jobs.submitShellCommandJob,Jobs.getJobStatus,Jobs.getDockerComputeDomains,Jobs.cancelJob

Examples

jobs = Jobs.getJobsList(top=2)

[Package SciServer version 2.0.0 Index]