Jobs.getRDBComputeDomainFromName {SciServer}	R Documentation

Get an RDBComputeDomain

Description

Returns an RDBComputeDomain object, given its registered name.

Usage

Jobs.getDockerComputeDomainsNames(rdbComputeDomainName, rdbComputeDomains = NULL)

Arguments

rdbComputeDomainName
name of the RDBComputeDomain, as shown within the results of Jobs.getRDBComputeDomainsNames() or Jobs.getRDBComputeDomains()

rdbComputeDomains
a list of RDBComputeDomain objects, as returned by Jobs.getRDBComputeDomains(). If not set, then an internal call to Jobs.getRDBComputeDomains() is made.

Value

an RDBComputeDomain object (list) that defines an RDB compute domain. A list of these kind of objects available to the user is returned by the function Jobs.getRDBComputeDomains().

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://www.github.com/sciserver/SciScript-R

See Also

Jobs.getRDBComputeDomains.

Examples

rdbComputeDomain = Jobs.getRDBComputeDomainFromName("rdbComputeDomainAtJHU");

[Package SciServer version 2.0.0 Index]