Jobs.getRDBComputeDomainsNames {SciServer}	R Documentation

Get database domains names.

Description

Returns the names of the RDB compute domains available to the user.

Usage

Jobs.getRDBComputeDomainsNames(rdbComputeDomains=NULL)

Arguments

rdbComputeDomains
a list of rdbComputeDomain objects, as returned by Jobs.getRDBComputeDomains(). If not set, then an extra internal call to Jobs.getRDBComputeDomains() is made.

Value

a list of strings, each being the name of a rdb compute domain available to the user.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://www.github.com/sciserver/SciScript-R

See Also

Jobs.getRDBComputeDomains.

Examples

rdbComputeDomainsNames = Jobs.getRDBComputeDomainsNames()

[Package SciServer version 2.0.0 Index]