Jobs.submitRDBQueryJob {SciServer}	R Documentation

Submit SQL query job

Description

Submits a sql query for execution (as an asynchronous job) inside a relational database (RDB) compute domain.

Usage

Jobs.submitNotebookJob(sqlQuery, rdbComputeDomain=NULL, databaseContextName = NULL, resultsName='queryResults', resultsFolderPath="", jobAlias = "")

Arguments

sqlQuery
sql query (string)

rdbComputeDomain
object (list) that defines a relational database (RDB) compute domain. A list of these kind of objects available to the user is returned by the function Jobs.getRDBComputeDomains().

databaseContextName
database context name (string) on which the sql query is executed.

resultsName
name (string) of the table or file (without file type ending) that contains the query result. In case the sql query has multiple statements, should be set to a list of names (e.g., ['result1','result2']).

dataVolumes
a list with the names of data volumes that will be mounted to the docker Image. E.g.: dataVolumes=list(list(name='SDSS DAS')). A list of available data volumes can be found as the 'volumes' property in the dockerComputeDomain list. If dataVolumes=NULL, then all available data volumes are mounted.

resultsFolderPath
full path to results folder (string) where query output tables are written into. E.g.: "/home/idies/workspace/rootVOlume/username/userVolume/jobsFolder"" . If not set, then a default folder will be set automatically.

jobAlias
alias (string) of job, defined by the user.

Value

the job ID (integer)

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://www.github.com/sciserver/SciScript-R

See Also

Jobs.submitNotebookJob,Jobs.submitShellCommandJob,Jobs.getJobStatus,Jobs.getRDBComputeDomains,Jobs.cancelJob

Examples

jobId = Jobs.submitRDBQueryJob("select 1", NULL, NULL, "myQueryResults", "myNewJob")

[Package SciServer version 2.0.0 Index]