Jobs.waitForJob {SciServer}	R Documentation

Wait for job status.

Description

Queries regularly the job status and waits until the job is completed.

Usage

Jobs.waitForJob(jobId, verbose=TRUE, pollTime = 5)

Arguments

jobId
id of job (integer).

verbose
if TRUE, will print "wait" messages on the screen while the job is still running. If FALSE, will suppress the printing of messages on the screen.

pollTime
idle time interval (integer, in seconds) before querying again for the job status. Minimum value allowed is 5 seconds.

Value

Returns a list object containing the job definition after the job is finished.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://apps.sciserver.org
http://www.github.com/sciserver/SciScript-R

See Also

Jobs.submitShellCommandJob,Jobs.getJobStatus,Jobs.getDockerComputeDomains

Examples

Jobs.cancelJob(123)

[Package SciServer version 2.0.0 Index]