LoginPortal.setToken {SciServer}	R Documentation

Set token (DEPRECATED).

Description

DEPRECATED. Use Authentication.setToken instead.

Sets the SciServer LoginPortal token of the user as the Sys environmental variable, as well as in the variable LoginPortal.token.

Usage

LoginPortal.setToken(token)

Arguments

token
Sciserver's LoginPortal token for the user

Author(s)

Gerard Lemson, Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://portal.sciserver.org/login-portal/
http://www.github.com/sciserver/SciScript-R

See Also

LoginPortal.getKeystoneUserWithToken, LoginPortal.setToken, LoginPortal.login, LoginPortal.getToken, Authentication.getKeystoneUserWithToken, Authentication.login, Authentication.getToken, Authentication.setToken, Authentication.token.

Examples

## Not run: 
LoginPortal.setToken("MyToken")
## End(Not run)

[Package SciServer version 2.0.0 Index]