SciDrive.upload {SciServer}	R Documentation

Upload file or data.

Description

Uploads data or a local file into a SciDrive directory.

Usage

SciDrive.upload(path, data="", localFilePath="")

Arguments

path
desired file path in SciDrive (string).

data
data to be uploaded into SciDrive. If the 'localFilePath' parameter is set, then the local file will be uploaded instead.

localFilePath
path to the local file to be uploaded (string).

Value

Returns an object with the attributes of the uploaded file. Throws an exception if the user is not logged into SciServer (use Authentication.login for that purpose). Throws an exception if the HTTP request to the SciDrive API returns an error.

Author(s)

Gerard Lemson, Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://www.scidrive.org/scidrive
http://www.github.com/sciserver/SciScript-R

See Also

SciDrive.createContainer.

Examples

## Not run: 
response = SciDrive.upload("/SciDrive/path/to/file.csv", "/local/path/to/file.csv")

## End(Not run)

[Package SciServer version 2.0.0 Index]