SciServer {SciServer}	R Documentation

Access to SciServer tools.

Description

This package provides functions for quick access of SciServer web services and tools.
SciServer (www.sciserver.org) provides a new online framework for data-intensive scientifc computing in the cloud, where the motto is to bring the computation close where the data is stored, and allow seamless access and sharing of big data sets within the scientific community.

Details

Some SciServer tools you can access with this package:

CasJobs: Database storage and querying.
You can have access big databases, as well as save your data tables in your own database called 'MyDB'. The user can run synchronous or asynchronous SQL queries and get the result back as an R data-frame (using CasJobs.executeQuery or CasJobs.submitJob, respectively). Uploading of CSV files or R data-frames into a database table can be done using CasJobs.uploadCSVDataToTable and CasJobs.uploadDataFrameToTable, respectively.

Login Portal: Single sign-on portal to all SciServer applications.
Although some tools accept anonymous access, you can use Authentication.login to login and access your own data and environment (after registering in the Login Portal).

SciDrive: Drag-and-drop file storage and sharing.
You can create directories in SciDrive using SciDrive.createContainer, upload a file to SciDrive using SciDrive.upload, and share its URL with your collaborators by using SciDrive.publicUrl.

SkyServer: Access to the SDSS astronomical survey.
You can query the SDSS database using SkyServer.sqlSearch, get cutout images from the sky using SkyServer.getJpegImgCutout.

SkyQuery: Access to the SkyQuery astronomical database catalog cross-match.
You can upload your own astronomical source catalog into your own database using SkyQuery.uploadTable, and cross-match it againt many other public source catalogs by running asynchronous SQL queries jobs using SkyQuery.submitJob.

Compute Jobs: Submission of Jupyter notebooks or shell commands as jobs.
You can execute whole Jupyter notebooks and shell commands as asynchronous batch jobs, as well as synchronous jobs, by using Jobs.submitShellCommandJob and Jobs.submitNotebookJob.

SciQuery Jobs: Submission of SQL queries.
You can execute synchronous or asynchronous SQL queries against Postgres and other database backends, by using Jobs.submitRDBQueryJob.

Files: Interaction with the SciServer file system.
You can upload and download data into your own file space in SciServer (e.g., Files.upload, Files.download

Package: 	 SciServer

Type: 	 Package

Version: 	 2.0.0

Author(s)

Manuchehr Taghizadeh-Popp, Gerard Lemson
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://www.github.com/sciserver/SciScript-R

[Package SciServer version 2.0.0 Index]