SkyQuery.cancelJob {SciServer}	R Documentation

Cancel a job

Description

Cancels a job already submitted (more info in http://www.voservices.net/skyquery).

Usage

SkyQuery.cancelJob(jobId)

Arguments

jobId
the ID of the job (string), which is obtained at the moment of submitting the job.

Value

Returns TRUE if the job was canceled successfully. Throws an exception if the user is not logged into SciServer (use Authentication.login for that purpose). Throws an exception if the HTTP request to the SkyQuery API returns an error.

Author(s)

Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://www.voservices.net/skyquery
http://www.github.com/sciserver/SciScript-R

See Also

SkyQuery.submitJob, SkyQuery.getJobStatus.

Examples

isCanceled = SkyQuery.cancelJob(SkyQuery.submitJob("select 1 as foo"))

[Package SciServer version 1.10.0 Index]