SkyQuery.getQueueInfo {SciServer}	R Documentation

Get queue info

Description

Returns information about a particular job queue (more info in http://www.voservices.net/skyquery).

Usage

SkyQuery.getQueueInfo(queue)

Arguments

queue
queue name (string).

Value

Returns a list containing information associated to the queue. Throws an exception if the user is not logged into SciServer (use Authentication.login for that purpose). Throws an exception if the HTTP request to the SkyQuery API returns an error.

Author(s)

Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://www.voservices.net/skyquery
http://www.github.com/sciserver/SciScript-R

See Also

SkyQuery.listQueues, SkyQuery.submitJob, SkyQuery.getJobStatus.

Examples

queueInfo = SkyQuery.getQueueInfo('quick')

[Package SciServer version 1.10.0 Index]