SkyQuery.listDatasetTables {SciServer}	R Documentation

List dataset tables.

Description

Returns a list of all tables within a dataset (more info in http://www.voservices.net/skyquery).

Usage

SkyQuery.listDatasetTables(datasetName="MyDB")

Arguments

datasetName
name of dataset (string).

Value

Returns a list containing the tables and associated descriptions/metadata. Throws an exception if the user is not logged into SciServer (use Authentication.login for that purpose). Throws an exception if the HTTP request to the SkyQuery API returns an error.

Author(s)

Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://www.voservices.net/skyquery
http://www.github.com/sciserver/SciScript-R

See Also

SkyQuery.listQueues, SkyQuery.listAllDatasets, SkyQuery.getDatasetInfo, SkyQuery.listDatasetTables, SkyQuery.getTableInfo, SkyQuery.listTableColumns, SkyQuery.getTable, SkyQuery.dropTable

Examples

tables = SkyQuery.listDatasetTables("MyDB")

[Package SciServer version 1.10.0 Index]