SkyQuery.listQueues {SciServer}	R Documentation

List queues

Description

Returns a list of all available job queues and related metadata (more info in http://www.voservices.net/skyquery).

Usage

SkyQuery.listQueues()

Value

Returns a list of all available job queues and related metadata (more info in http://www.voservices.net/skyquery). Throws an exception if the user is not logged into SciServer (use Authentication.login for that purpose). Throws an exception if the HTTP request to the SkyQuery API returns an error.

Author(s)

Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://www.voservices.net/skyquery
http://www.github.com/sciserver/SciScript-R

See Also

SkyQuery.getQueueInfo, SkyQuery.submitJob, SkyQuery.getJobStatus.

Examples

queueList = SkyQuery.listQueues()

[Package SciServer version 2.0.0 Index]