SkyQuery.waitForJob {SciServer}	R Documentation

Wait for finished job.

Description

Queries the job status from SkyQuery every 2 seconds and waits for the SkyQuery job to be completed.

Usage

SkyQuery.waitForJob(jobId, verbose=TRUE)

Arguments

jobId
id of job (string).

verbose
message printing option.

If TRUE, will print "wait" messages on the screen while the job is not done. If FALSE, will suppress printing messages on the screen.

Value

Returns a list containing the job status and related metadata. Throws an exception if the user is not logged into SciServer (use Authentication.login for that purpose). Throws an exception if the HTTP request to the SkyQuery API returns an error.

Author(s)

Manuchehr Taghizadeh-Popp
Maintainer: Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://skyserver.sdss.org/casjobs
http://www.github.com/sciserver/SciScript-R

See Also

SkyQuery.submitJob, SkyQuery.getJobStatus.

Examples

SkyQuery.waitForJob(SkyQuery.submitJob("select 1"))

[Package SciServer version 2.0.0 Index]