SkyServer.getJpegImgCutout {SciServer}	R Documentation

Get a sky image cutout from SDSS.

Description

Gets a rectangular image cutout from a region of the sky in SDSS, centered at (ra,dec).

Usage

SkyServer.getJpegImgCutout(ra, dec, scale=0.7, width=512, height=512, opt="", query="", dataRelease=NULL)

Arguments

ra
Right Ascension of the image's center.

dec
Declination of the image's center.

scale
Scale of the image, measured in [arcsec/pix]

width
Width of the image, measured in [pix]

height
Height of the image, measured in [pix]

opt
Optional drawing options, expressed as concatenation of letters (string). The letters options are
"G": Grid. Draw a N-S E-W grid through the center
"L": Label. Draw the name, scale, ra, and dec on image.
"P PhotoObj. Draw a small cicle around each primary photoObj.
"S: SpecObj. Draw a small square around each specObj.
"O": Outline. Draw the outline of each photoObj.
"B": Bounding Box. Draw the bounding box of each photoObj.
"F": Fields. Draw the outline of each field.
"M": Masks. Draw the outline of each mask considered to be important.
"Q": Plates. Draw the outline of each plate.
"I": Invert. Invert the image (B on W).
(see http://skyserver.sdss.org/public/en/tools/chart/chartinfo.aspx)

query
Optional string. Marks with inverted triangles on the image the position of user defined objects. The (RA,Dec) coordinates of these object can be given by three means:
1) query is a SQL command of format "SELECT Id, RA, Dec, FROM Table".
2) query is list of objects. A header with RA and DEC columns must be included. Columns must be separated by tabs, spaces, commas or semicolons. The list may contain as many columns as wished.
3) aquery is a string following the pattern: ObjType Band (low_mag, high_mag).
ObjType: S | G | P marks Stars, Galaxies or PhotoPrimary objects.
Band: U | G | R | I | Z | A restricts marks to objects with Band BETWEEN low_mag AND high_mag Band 'A' will mark all objects within the specified magnitude range in any band (ORs composition).
Examples:
S
S R (0.0, 23.5)
G A (20, 30)
(see http://skyserver.sdss.org/public/en/tools/chart/chartinfo.aspx)

dataRelease
SDSS data release. E.g, 'DR13'

Value

Returns an array of doubles, of dimensions [1:width,1:height,1:3]. Throws an exception if the HTTP request to the SkyServer API returns an error.

Author(s)

Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://skyserver.sdss.org
http://www.github.com/sciserver/SciScript-R

Examples

## Not run: 
img = SkyServer.getJpegImgCutout(ra=197.614455642896, dec=18.438168853724, width=512, height=512, scale=0.4, opt="OG", query="SELECT TOP 100 p.objID, p.ra, p.dec, p.r FROM fGetObjFromRectEq(197.6,18.4,197.7,18.5) n, PhotoPrimary p WHERE n.objID=p.objID")
plot(0:1, 0:1, type = "n");
rasterImage(img, 0, 0, 1, 1)

## End(Not run)

[Package SciServer version 1.10.0 Index]