SkyServer.objectSearch {SciServer}	R Documentation

Searches for an object in the SDSS catalog.

Description

Searches for an object in the SDSS catalog and gets its properties and associated metadata. Search parameters include position in the sky and various identifiers included in SDSS catalog tables.

Usage

SkyServer.objectSearch(objId=NULL, specObjId=NULL, apogee_id=NULL, apstar_id=NULL, ra=NULL, dec=NULL, plate=NULL, mjd=NULL, fiber=NULL, run=NULL, rerun=NULL, camcol=NULL, field=NULL, obj=NULL, dataRelease=NULL)

Arguments

objId
SDSS ObjId.

specObjId
SDSS SpecObjId.

apogee_id
ID identifying Apogee target object.

apstar_id
unique ID for combined apogee star spectrum.

ra
Right Ascension of the object.

dec
Declination of the object.

plate
SDSS plate number.

mjd
Modified Julian Date of observation.

fiber
SDSS fiber number.

run
SDSS run number.

rerun
SDSS rerun number.

camcol
SDSS camera column.

field
SDSS field number.

obj
The object id within a field.

dataRelease
SDSS data release. E.g, 'DR13'

Value

Returns an object containing various tables with the objects properties and associated metadata. Throws an exception if the HTTP request to the SkyServer API returns an error.

Author(s)

Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://skyserver.sdss.org
http://www.github.com/sciserver/SciScript-R

See Also

SkyServer.sqlSearch, SkyServer.rectangularSearch, SkyServer.radialSearch.

Examples

object = SkyServer.objectSearch(ra=258.25, dec=64.05)

[Package SciServer version 1.10.0 Index]