SkyServer.radialSearch {SciServer}	R Documentation

Run a radial search on the sky.

Description

Runs a query in the SDSS database that searches for all objects within a certain radius from a point in the sky, and retrieves the result table as a dataframe. Maximum number of rows retrieved is set currently to 500,000.

Usage

SkyServer.radialSearch(ra, dec, radius=1, coordType="equatorial", whichPhotometry="optical", limit="0", dataRelease=NULL)

Arguments

ra
Right Ascension of the point in the sky.

dec
Declination of the point in the sky.

radius
Search radius around the (ra,dec) coordinate in the sky. Measured in arcminutes.

coordType
Type of celestial coordinate system. Can be set to "equatorial" or "galactic".

whichPhotometry
Type of retrieved data. Can be set to "optical" or "infrared".

limit
maximum number of rows in the result table (string). If set to "0", then the function will return all rows.

dataRelease
SDSS data release. E.g, 'DR13'

Value

Returns the results table as a dataframe. Throws an exception if the HTTP request to the SkyServer API returns an error.

Author(s)

Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://skyserver.sdss.org
http://www.github.com/sciserver/SciScript-R

See Also

SkyServer.sqlSearch, SkyServer.rectangularSearch.

Examples

table = SkyServer.radialSearch(ra=258.25, dec=64.05, radius=3)

[Package SciServer version 1.10.0 Index]