SkyServer.rectangularSearch {SciServer}	R Documentation

Run a rectangular search on the sky.

Description

Runs a query in the SDSS database that searches for all objects within a certain rectangular box defined on the the sky, and retrieves the result table as a dataframe. Maximum number of rows retrieved is set currently to 500,000.

Usage

SkyServer.rectangularSearch(min_ra, max_ra, min_dec, max_dec, coordType="equatorial", whichPhotometry="optical", limit="0", dataRelease=NULL)

Arguments

min_ra
Minimum value of Right Ascension coordinate that defines the box boundaries on the sky.

max_ra
Maximum value of Right Ascension coordinate that defines the box boundaries on the sky.

min_dec
Minimum value of Declination coordinate that defines the box boundaries on the sky.

max_dec
Maximum value of Declination coordinate that defines the box boundaries on the sky.

coordType
Type of celestial coordinate system. Can be set to "equatorial" or "galactic".

whichPhotometry
Type of retrieved data. Can be set to "optical" or "infrared".

limit
Maximum number of rows in the result table (string). If set to "0", then the function will return all rows.

dataRelease
SDSS data release. E.g, 'DR13'

Value

Returns the results table as a dataframe. Throws an exception if the HTTP request to the SkyServer API returns an error.

Author(s)

Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://skyserver.sdss.org
http://www.github.com/sciserver/SciScript-R

See Also

SkyServer.sqlSearch, SkyServer.radialSearch.

Examples

table = SkyServer.rectangularSearch(min_ra=258.2, max_ra=258.3, min_dec=64,max_dec=64.1)

[Package SciServer version 2.0.0 Index]