SkyServer.sqlSearch {SciServer}	R Documentation

Execute SQL query.

Description

Executes a SQL query to the SDSS database, and retrieves the result table as a dataframe. Maximum number of rows retrieved is set currently to 500,000.

Usage

SkyServer.sqlSearch(sql, dataRelease=NULL)

Arguments

sql
a string containing the sql query

dataRelease
SDSS data release. E.g, 'DR13'. Default value already set in Config.DataRelease

Value

Returns the results table as a dataframe. Throws an exception if the HTTP request to the SkyServer API returns an error.

Author(s)

Manuchehr Taghizadeh-Popp <mtaghiza@jhu.edu>

References

http://www.sciserver.org
http://skyserver.sdss.org
http://www.github.com/sciserver/SciScript-R

See Also

CasJobs.executeQuery, CasJobs.submitJob.

Examples

df = SkyServer.sqlSearch(sql="select 1")

[Package SciServer version 2.0.0 Index]